distro.sh: distribute package file without zip wrapper
I've now configured the web servers (both nginx and Apache) to server
".pkg" files with a MIME type of "application/octet-stream" instead of
the default "text/plain". This means that it is no longer necessary to
wrap them up inside zip archives to prevent them from being
inappropriately displayed in the user's browser window rather than
downloaded to disk.

So, drop the zip wrapper and instead just give the package a unique
name. Flat packages are compressed in any case, so the download size
should be basically the same, give or take a few bytes.

Signed-off-by: Wincent Colaiuta <win@wincent.com>